God came from heavens and gave me to power and will to do this, Praise the Ommnissiah for this achievement, now be free my children, explore a limitless world generated by the power of advanced compuation maths until they start crashing in 3 billion of blocks away from the spawn now BEHOLD THE OPEN(Buggy)WORLD MODE, i

+++ core/src/io/anuke/mindustry/core/NetClient.java
@@ -290,7 +290,7 @@ public class NetClient extends Module{
 
         byte cores = input.readByte();
         for(int i = 0; i < cores; i++){
-            int pos = input.readInt();
+            long pos = input.readLong();
             Tile tile = world.tile(pos);
             if(tile != null && tile.entity != null){
                 tile.entity.items.read(input);
